+++ src/domain/search/domain/src/repos/search_repository.rs
@@ -41,6 +41,8 @@ pub trait SearchRepository: Send + Sync {
         operations: Vec<SearchIndexUpdateOperation>,
     ) -> Result<(), InternalError>;
 
+    async fn drop_schema(&self, schema_name: SearchEntitySchemaName) -> Result<(), InternalError>;
+
     async fn drop_all_schemas(&self) -> Result<(), InternalError>;
 
     async fn listing_search(