+++ src/infra/flow-system/sqlite/tests/tests/test_sqlite_flow_process_state.rs
@@ -122,6 +122,69 @@ database_transactional_test!(
 database_transactional_test!(
     storage = sqlite,
     fixture =
+        kamu_flow_system_repo_tests::test_flow_process_state::test_rollup_from_csv_filtered_by_last_attempt_between,
+    harness = SqliteFlowProcessStateHarness
+);
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+database_transactional_test!(
+    storage = sqlite,
+    fixture =
+        kamu_flow_system_repo_tests::test_flow_process_state::test_rollup_from_csv_filtered_by_last_failure_since,
+    harness = SqliteFlowProcessStateHarness
+);
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+database_transactional_test!(
+    storage = sqlite,
+    fixture =
+        kamu_flow_system_repo_tests::test_flow_process_state::test_rollup_from_csv_filtered_by_next_planned_before,
+    harness = SqliteFlowProcessStateHarness
+);
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+database_transactional_test!(
+    storage = sqlite,
+    fixture =
+        kamu_flow_system_repo_tests::test_flow_process_state::test_rollup_from_csv_filtered_by_next_planned_after,
+    harness = SqliteFlowProcessStateHarness
+);
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+database_transactional_test!(
+    storage = sqlite,
+    fixture =
+        kamu_flow_system_repo_tests::test_flow_process_state::test_rollup_from_csv_filtered_by_min_consecutive_failures,
+    harness = SqliteFlowProcessStateHarness
+);
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+database_transactional_test!(
+    storage = sqlite,
+    fixture =
+        kamu_flow_system_repo_tests::test_flow_process_state::test_rollup_from_csv_combined_time_filters,
+    harness = SqliteFlowProcessStateHarness
+);
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+database_transactional_test!(
+    storage = sqlite,
+    fixture =
+        kamu_flow_system_repo_tests::test_flow_process_state::test_rollup_from_csv_combined_failures_and_time,
+    harness = SqliteFlowProcessStateHarness
+);
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+database_transactional_test!(
+    storage = sqlite,
+    fixture =
         kamu_flow_system_repo_tests::test_flow_process_state::test_list_processes_from_csv_unfiltered_with_default_ordering,
     harness = SqliteFlowProcessStateHarness
 );